Improve the approximation of power from current
P = Integral from Q1 to Q2 of V(Q)dQ

Use:

 (V1 + V2)/2 * (Q2 - Q1)

as the trapezoid rule approximation of this instead of:

 V2 * Q2 - V1 * Q1

which doesn't have a clear meaning. If V is constant, both are
the same, but for example, if we had a capacitor with V(Q) = Q/C,
then the old expression would over-estimate by a factor of 2,
while the new expression is asymptotically right for small changes.

https://bugzilla.gnome.org/show_bug.cgi?id=747370