Use inhibitor lock to ensure we can write the log
We will acquire a delay lock from logind, immediately write the
log file when we get the sleep notification, and the write it
again when we get the callback from the runner that the test has
stopped. This is also when we release the lock. Since we 'only'
acquire a delay lock, not a block lock, we might actually not
finish the test in time.